another workaround for an issue if dsyev eigenvector signs differ between processes
When the small eigenvalue problem is solved in solve_H_d.c, the signs of
the resulting eigenvectors are arbitrary. This is in itself not an
issue, as the sign of an eigenvector is not well defined: if v is an
eigenvector, so is -v. However, when running in parallel the signs of
the eigenvectors can differ between the different processes. (This does
not seem to happen on most platforms and LAPACK libraries, but we have
observed the issue on Win32 with the MKL library: Diagonalizing the same
matrix gives eigenvectors with different signs for the different
processes running primme in parallel.)

This is not a problem during the calculation, but if the complete
eigenvectors are assembled in the end from the parts of the basis
residing on the different processes, the sign of each part is
essentially random, leading to wrong results.

This commit implements a workaround, where after the diagonalization of
the small eigenvalue problem the eigenvectors are transferred from the
process with procID=0 to all other processes, thus ensuring a consistent
sign everywhere. This is done using the globalSumDouble function with a
zero vector as sendBuf on the receiving processes and the eigenvector on
procID=0. Because sendBuf and recvBuf should not overlap for MPI, rwork
is abused as a scratch vector. Due to the limited size of rwork, this
has to be done for every eigenvector individually.

NOTE: The workaround was only implemented for diagonalization of real
matrices. It should of course work in the same way for complex matrices,
but this was neither implemented nor tested ...